requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
opencv-python>=4.5.0
numpy>=1.20.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
flask-sqlalchemy>=3.0.0
sqlalchemy>=1.4.0
itsdangerous>=2.0.0
email-validator>=1.1.0 